+++ src/admin-hub/AdminPage.tsx
@@ -16,6 +16,7 @@ import { Surface, SurfaceBackground } from 'azure-devops-ui/Surface';
 import { ZeroData, ZeroDataActionType } from 'azure-devops-ui/ZeroData';
 import React, { useEffect, useMemo, useState } from 'react';
 
+import { groupBy } from '../common/helpers';
 import AddRuleResult from '../common/models/AddRuleResult';
 import Rule from '../common/models/Rule';
 import RuleDocument from '../common/models/WorkItemRules';
@@ -25,7 +26,7 @@ import WorkItemService from '../common/services/WorkItemService';
 import webLogger from '../common/webLogger';
 import LoadingSection from '../shared-ui/component/LoadingSection';
 import WorkItemTypeTag from '../shared-ui/component/WorkItemTypeTag';
-import { getCommandBarItems, getListColumns, groupBy, isGroup } from './helpers';
+import { getCommandBarItems, getListColumns, isGroup } from './helpers';
 
 const AdminPage = (): React.ReactElement => {
   const [types, setTypes] = useState<WorkItemType[]>([]);
