+++ src/admin-hub/helpers.tsx
@@ -234,18 +234,5 @@ export const getCommandBarItems = (
     }
   }
 ];
-function groupBy<T>(list: T[], keyGetter: (value: T) => string): Map<string, T[]> {
-  const map = new Map<string, T[]>();
-  list.forEach(item => {
-    const key = keyGetter(item);
-    const collection = map.get(key);
-    if (!collection) {
-      map.set(key, [item]);
-    } else {
-      collection.push(item);
-    }
-  });
-  return map;
-}
 
-export { getState, getWorkItemType, isGroup, groupBy };
+export { getState, getWorkItemType, isGroup };
